delete cache when segment released
