+++ internal/core/src/segcore/ChunkedSegmentSealedImpl.cpp
@@ -27,6 +27,7 @@
 
 #include "Utils.h"
 #include "Types.h"
+#include "common/GeometryCache.h"
 #include "common/Array.h"
 #include "common/Chunk.h"
 #include "common/ChunkWriter.h"
@@ -1309,6 +1310,10 @@ ChunkedSegmentSealedImpl::ChunkedSegmentSealedImpl(
 }
 
 ChunkedSegmentSealedImpl::~ChunkedSegmentSealedImpl() {
+    // Clean up geometry cache for all fields in this segment
+    auto& cache_manager = milvus::exec::SimpleGeometryCacheManager::Instance();
+    cache_manager.RemoveSegmentCaches(this);
+
     auto cc = storage::MmapManager::GetInstance().GetChunkCache();
     if (cc == nullptr) {
         return;
