+++ internal/core/src/segcore/segment_c.cpp
@@ -35,6 +35,7 @@
 #include "futures/Future.h"
 #include "futures/Executor.h"
 #include "exec/expression/ExprCache.h"
+#include "common/GeometryCache.h"
 
 //////////////////////////////    common interfaces    //////////////////////////////
 CStatus
@@ -90,6 +91,11 @@ NewSegment(CCollection collection,
 void
 DeleteSegment(CSegmentInterface c_segment) {
     auto s = static_cast<milvus::segcore::SegmentInterface*>(c_segment);
+
+    // Clean up geometry cache for all fields in this segment
+    auto& cache_manager = milvus::exec::SimpleGeometryCacheManager::Instance();
+    cache_manager.RemoveSegmentCaches(s);
+
     delete s;
 }
